# siemagent/internal/api/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: SIEMAgent API
  version: "1.0.0"
  description: |
    AI-powered Security Information and Event Management (SIEM) classifier.

    Parses syslog / JSON logs, evaluates Sigma detection rules, classifies
    events with an LLM (Gemini, Ollama or any OpenAI-compatible API), indexes
    them in Qdrant for semantic search, and exposes analytics.

    **Authentication:** when the server runs with `SIEM_API_KEYS`, every route
    except `/health`, `/health/ready` and `/docs` needs `Authorization: Bearer <key>`
    or `X-API-Key: <key>`.

    **Severity scale**
    | Code | Meaning |
    |------|---------|
    | P1 | Critical — active breach / ransomware / data exfil |
    | P2 | High — successful intrusion / privilege escalation |
    | P3 | Medium — failed attack attempt / anomaly |
    | P4 | Low — policy violation / reconnaissance |
    | P5 | Info — normal ops / noise |
servers:
  - url: http://localhost:8080
    description: Local dev server
...
paths:
...
  /classify:
    post:
      tags: [Classification]
      summary: Classify a single log line (legacy top-level path)
      description: Alias for `POST /api/classify`. Kept for backward compatibility.
      operationId: classifyLegacy
      requestBody:
        $ref: "#/components/requestBodies/ClassifyRequest"
      responses:
        "200":
          $ref: "#/components/responses/ClassifiedEvent"
        "400":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /classify/stream:
    post:
      tags: [Classification]
      summary: Classify a log line with SSE streaming (legacy top-level path)
      description: Alias for `POST /api/classify/stream`. Kept for backward compatibility.
      operationId: classifyLogStreamLegacy
      requestBody:
        $ref: "#/components/requestBodies/ClassifyRequest"
      responses:
        "200":
          description: SSE stream — see `/api/classify/stream` for full event format
          content:
            text/event-stream:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/ValidationError"
  /api/classify:
    post:
      tags: [Classification]
      summary: Classify a single log line
      description: |
        Parses the raw log line (auto-detecting RFC 5424 syslog, RFC 3164 syslog,
        JSON, or treating it as a raw string), then sends it to the LLM for classification.

        **Validation rules**
        - `log` is required, max 8 192 bytes
        - `format` must be `syslog`, `json`, or `auto` (default `auto`)
      operationId: classifyLog
      requestBody:
        $ref: "#/components/requestBodies/ClassifyRequest"
      responses:
        "200":
          $ref: "#/components/responses/ClassifiedEvent"
        "400":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/classify/stream:
    post:
      tags: [Classification]
      summary: Classify a log line with SSE streaming
      description: |
        Same as `POST /api/classify` but streams the LLM response as
        Server-Sent Events so the client receives incremental text.

        **Event types**
        ```
        data: {"chunk":"..."}            ← incremental LLM token
        data: {"result":{...},"done":true}  ← final ClassifiedEvent + done signal
        data: {"error":"..."}           ← classification failure
        ```

        Set `Accept: text/event-stream` or let the browser handle it via `EventSource`.
      operationId: classifyLogStream
      requestBody:
        $ref: "#/components/requestBodies/ClassifyRequest"
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string
              example: |
                data: {"chunk":"{\"attack_type\":\"Brute"}

                data: {"chunk":" Force\",\"severity\":\"P2\"..."}

                data: {"result":{"attack_type":"Brute Force","severity":"P2","confidence":0.95,"iocs":["192.168.1.100"],"summary":"Multiple failed SSH login attempts.","remediation":"Block the source IP and rotate credentials.","mitre":{"tactic":"Credential Access","technique_id":"T1110.001","technique":"Password Guessing"},"event":{"raw":"<165>1 2024-01-15T10:30:00Z webserver sshd 1234 - Failed password for root from 192.168.1.100 port 22","timestamp":"2024-01-15T10:30:00Z","hostname":"webserver","app_name":"sshd","proc_id":"1234","message":"Failed password for root from 192.168.1.100 port 22","source":"syslog"},"processed_at":"2024-01-15T10:30:01Z"},"done":true}

        "400":
          $ref: "#/components/responses/ValidationError"

  # ── Ingestion ─────────────────────────────────────────────────────────────────
  /api/ingest:
    post:
      tags: [Ingestion]
      summary: Bulk ingest up to 500 log lines
      description: |
        Parses and classifies all lines using the worker pool.
        Streams chunked progress lines every 10 events, then a final summary object.

        **Progress line format (chunked)**
        ```json
        {"progress":10,"accepted":47}
        ```
        **Final response object**
        ```json
        {"accepted":47,"classified":45,"errors":2,"results":[...]}
        ```
      operationId: ingestLogs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestRequest"
            examples:
              syslog_batch:
                summary: Batch of syslog lines
                value:
                  format: syslog
                  logs:
                    - "<165>1 2024-01-15T10:30:00Z webserver sshd 1234 - Failed password for root from 192.168.1.100 port 22"
                    - "<165>1 2024-01-15T10:30:01Z webserver sshd 1234 - Failed password for root from 192.168.1.100 port 22"
                    - "<165>1 2024-01-15T10:30:02Z webserver sshd 1234 - Accepted publickey for deploy from 10.0.0.5 port 41222"
              json_batch:
                summary: Batch of JSON log lines
                value:
                  format: json
                  logs:
                    - '{"level":"error","msg":"auth failed","time":"2024-01-15T10:30:00Z","hostname":"api-01"}'
                    - '{"level":"warn","msg":"rate limit exceeded","time":"2024-01-15T10:30:05Z","hostname":"api-01"}'
      responses:
        "200":
          description: |
            Bulk classification result streamed as chunked JSON.

            The response uses `Transfer-Encoding: chunked`. Every 10 events a
            progress line is flushed, followed by the final IngestResponse object:

            ```
            {"progress":10,"accepted":47}
            {"progress":20,"accepted":47}
            {"accepted":47,"classified":45,"errors":2,"results":[...]}
            ```
          headers:
            Transfer-Encoding:
              schema:
                type: string
                enum: [chunked]
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResponse"
              example:
                accepted: 3
                classified: 3
                errors: 0
                results:
                  - attack_type: Brute Force
                    severity: P2
                    confidence: 0.95
                    iocs: ["192.168.1.100", "root"]
                    summary: Multiple failed SSH login attempts from 192.168.1.100.
                    remediation: Block the source IP and reset account credentials.
                    mitre:
                      tactic: Credential Access
                      technique_id: T1110.001
                      technique: "Password Guessing"
                    event:
                      raw: "<165>1 2024-01-15T10:30:00Z webserver sshd 1234 - Failed password for root from 192.168.1.100 port 22"
                      timestamp: "2024-01-15T10:30:00Z"
                      hostname: webserver
                      app_name: sshd
                      proc_id: "1234"
                      message: Failed password for root from 192.168.1.100 port 22
                      source: syslog
                    processed_at: "2024-01-15T10:30:01Z"
        "400":
          description: Request validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                empty_logs:
                  summary: Empty logs array
                  value:
                    error: "logs array is empty"
                too_many_logs:
                  summary: Exceeds 500 line limit
                  value:
                    error: "maximum 500 log lines per request"
                invalid_json:
                  summary: Malformed request body
                  value:
                    error: "invalid JSON body"

  # ── Search ────────────────────────────────────────────────────────────────────
  /api/search:
    get:
      tags: [Search]
      summary: Semantic search over indexed events
      description: |
        Embeds the query text with Ollama (nomic-embed-text) and performs a
        nearest-neighbour search in Qdrant.

        Returns 503 when Qdrant / Ollama are not configured.
      operationId: searchEvents
      parameters:
        - name: q
          in: query
          required: true
          description: Natural language search query
          schema:
            type: string
            minLength: 1
          example: "brute force SSH login from China"
        - name: severity
          in: query
          required: false
          description: Filter results to a single severity level
          schema:
            $ref: "#/components/schemas/Severity"
          example: P2
        - name: limit
          in: query
          required: false
          description: Maximum number of results (1–100, default 20)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          example: 10
      responses:
        "200":
          description: Ranked search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SearchHit"
              example:
                - event_id: "Failed password f"
                  timestamp: "2024-01-15T10:30:01Z"
                  source: syslog
                  attack_type: Brute Force
                  severity: P2
                  summary: Multiple failed SSH login attempts from 192.168.1.100.
                  mitre_tactic: Credential Access
                  score: 0.97
                - event_id: "Failed password f"
                  timestamp: "2024-01-14T08:12:44Z"
                  source: syslog
                  attack_type: Brute Force
                  severity: P2
                  summary: SSH brute force attempt targeting root account from 203.0.113.5.
                  mitre_tactic: Credential Access
                  score: 0.91
        "400":
          description: Missing required query parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "q parameter is required"
        "500":
          description: Embedding or vector search failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                embed_failed:
                  summary: Ollama embedding call failed
                  value:
                    error: "embedding failed"
                search_failed:
                  summary: Qdrant search call failed
                  value:
                    error: "search failed"
        "503":
          description: Semantic search not configured (Qdrant/Ollama not available)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "semantic search not configured (Qdrant/Ollama not available)"

  # ── Events and detection ──────────────────────────────────────────────────────
  /api/events:
    get:
      tags: [Events]
      summary: Recent classified events
      description: Most recent stored events, newest first. With Postgres this includes history from before the last restart.
      operationId: listEvents
      parameters:
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 500, default: 100}
      responses:
        "200":
          description: Events, newest first
          content:
            application/json:
              schema:
                type: array
                items: {$ref: "#/components/schemas/ClassifiedEvent"}
        "400":
          description: limit out of range
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ValidationError"}
...
  /api/detections/rules/{id}:
    patch:
      tags: [Detection]
      summary: Enable or disable a detection rule
      description: |
        Takes effect immediately and is saved (in Postgres when configured),
        so it survives restarts. Disabling a threshold rule clears its counters.
      operationId: updateDetectionRule
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled: {type: boolean}
      responses:
        "200":
          description: The updated rule
          content:
            application/json:
              schema: {$ref: "#/components/schemas/DetectionRule"}
        "400": {description: "Body is not {\"enabled\": true|false}"}
        "404": {description: Unknown rule, or detection is off}

  # ── Incidents ─────────────────────────────────────────────────────────────────
  /api/incidents:
    get:
      tags: [Incidents]
      summary: List incidents
      description: Most recently active first. Alerts sharing an IP, user or host within INCIDENT_WINDOW are grouped into one incident.
      operationId: listIncidents
      parameters:
        - {name: status, in: query, schema: {type: string, enum: [new, investigating, resolved]}}
        - {name: severity, in: query, schema: {type: string, enum: [P1, P2, P3, P4, P5]}}
        - {name: assignee, in: query, schema: {type: string}}
        - name: entity
          in: query
          description: Only incidents involving this entity
          schema: {type: string, example: "ip:203.0.113.7"}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 500, default: 100}}
      responses:
        "200":
          description: Incidents
          content:
            application/json:
              schema:
                type: array
                items: {$ref: "#/components/schemas/Incident"}
        "400": {description: Invalid entity or limit}
        "404": {description: Incidents are not enabled}
...
  /api/incidents/{id}:
    parameters:
      - {name: id, in: path, required: true, schema: {type: string, example: INC-2EA4C98D}}
    get:
      tags: [Incidents]
      summary: Incident with alerts and history
      operationId: getIncident
      responses:
        "200":
          description: The incident
          content:
            application/json:
              schema: {$ref: "#/components/schemas/IncidentDetail"}
        "404": {description: Unknown incident}
    patch:
      tags: [Incidents]
      summary: Update status, assignee, severity or resolution
      description: Every changed field is recorded in the incident history. A resolution needs status resolved; reopening clears it.
      operationId: updateIncident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status: {type: string, enum: [new, investigating, resolved]}
                assignee: {type: string, maxLength: 100}
                severity: {type: string, enum: [P1, P2, P3, P4, P5]}
                resolution: {type: string, enum: ["", true_positive, false_positive, benign, duplicate]}
      responses:
        "200":
          description: The updated incident
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Incident"}
        "400": {description: Invalid change}
        "404": {description: Unknown incident}
  /api/incidents/{id}/comments:
    post:
      tags: [Incidents]
      summary: Add a comment
      operationId: addIncidentComment
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: {type: string, maxLength: 5000}
      responses:
        "201":
          description: The recorded comment
          content:
            application/json:
              schema: {$ref: "#/components/schemas/IncidentActivity"}
        "400": {description: Empty or too long}
        "404": {description: Unknown incident}
  /api/incidents/{id}/investigate:
    post:
      tags: [Incidents]
      summary: Run an AI investigation of the whole incident
      description: |
        Runs in the background; the write-up streams over /ws/alerts and is
        saved to the incident history as an "investigation" entry.
      operationId: investigateIncident
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        "202": {description: Started}
        "404": {description: Unknown incident}
        "409": {description: An investigation is already running for this incident}
        "429": {description: Too many investigations in flight; retry later}
        "503": {description: No LLM agent configured}
  /api/incidents/{id}/report:
    get:
      tags: [Incidents]
      summary: Markdown incident report
      operationId: getIncidentReport
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
        - {name: download, in: query, description: Serve as an attachment, schema: {type: string, example: "1"}}
      responses:
        "200":
          description: The report
          content:
            text/markdown:
              schema: {type: string}
        "404": {description: Unknown incident}
...
  /api/suppressions/{id}:
    delete:
      tags: [Incidents]
      summary: Lift a suppression (analyst)
      operationId: liftSuppression
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        "204": {description: Lifted}
        "404": {description: Not found}
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "semantic search not configured"
...
  responses:
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
          examples:
            empty_log:
              summary: Missing log field
              value:
                error: "log field is required"
                field: log
            oversized:
              summary: Log line too large
              value:
                error: "log field must not exceed 8192 bytes"
                field: log
            bad_format:
              summary: Invalid format value
              value:
                error: "format must be one of: syslog, json, auto"
                field: format
...